improve snapshot restore to support nested snapshots
