+++ package/cloudshell/cp/vcenter/commands/command_orchestrator.py
@@ -10,7 +10,7 @@ from cloudshell.cp.vcenter.commands.power_manager_vm import VirtualMachinePowerM
 from cloudshell.cp.vcenter.commands.refresh_ip import RefreshIpCommand
 from cloudshell.cp.vcenter.commands.restore_snapshot import SnapshotRestorer
 from cloudshell.cp.vcenter.commands.save_snapshot import SaveSnapshotCommand
-from cloudshell.cp.vcenter.commands.snapshots_retriever import SnapshotRetriever
+from cloudshell.cp.vcenter.commands.snapshots_retriever import SnapshotRetriever, SnapshotRetrieverCommand
 from cloudshell.cp.vcenter.common.cloud_shell.driver_helper import CloudshellDriverHelper
 from cloudshell.cp.vcenter.common.cloud_shell.resource_remover import CloudshellResourceRemover
 from cloudshell.cp.vcenter.common.model_factory import ResourceModelParser
@@ -128,7 +128,7 @@ class CommandOrchestrator(object):
                                                   resource_model_parser=ResourceModelParser(),
                                                   task_waiter=synchronous_task_waiter)
 
-        self.snapshots_retriever = SnapshotRetriever(pyvmomi_service=pv_service)
+        self.snapshots_retriever = SnapshotRetrieverCommand(pyvmomi_service=pv_service)
 
     def connect_bulk(self, context, request):
         results = self.command_wrapper.execute_command_with_connection(
